refactor: Update utilities, add new ones
