+++ src/commands/gw2/account.js
@@ -1,7 +1,7 @@
 const { Keys, Worlds } = require("../../models");
 const { MessageEmbed } = require("discord.js");
 const { gw2api, getLeadingGuilds } = require("../../utils/api");
-const { formatAge, filterExpansions } = require("../../utils/general");
+const { formatAge, filterExpansions } = require("../../utils");
 
 class Account {
   constructor() {
