+++ src/commands/gw2/character.js
@@ -1,7 +1,7 @@
 const { Keys } = require("../../models");
 const { MessageEmbed } = require("discord.js");
 const { gw2api } = require("../../utils/api");
-const { formatAge, sortAlphabetically } = require("../../utils/general");
+const { formatAge, sortAlphabetically } = require("../../utils");
 const { professions } = require("../../utils/gameData");
 const logger = require("../../utils/logger");
