+++ src/commands/gw2/dailies.js
@@ -1,5 +1,5 @@
 const { MessageEmbed } = require("discord.js");
-const { sortAlphabetically } = require("../../utils/general");
+const { sortAlphabetically } = require("../../utils");
 const { gw2api } = require("../../utils/api");
 const { Achievements } = require("../../models");
 const { fractalsData } = require("../../utils/gameData");
